Rename Python module to requests
Plus restructure of the repo to start matching the official requests
module
