--- wasm_requests/__init__.py
@@ -1,1 +1,0 @@
-from .wasm_requests import *
